+++ _localtest/src/main/java/test/MyClass.java
@@ -1,4 +1,6 @@
 package test;
+import com.ctre.phoenix.motorcontrol.ControlMode;
+import com.ctre.phoenix.motorcontrol.can.TalonSRX;
 import com.kauailabs.navx.frc.AHRS;
 import edu.wpi.first.wpilibj.I2C;
 import edu.wpi.first.wpilibj.IterativeRobot;
@@ -7,19 +9,19 @@ import edu.wpi.first.wpilibj.Utility;
 public class MyClass extends IterativeRobot {
 
     AHRS ahrs;
-//    TalonSRX talonSrx;
+    TalonSRX talonSrx;
 
     public void robotInit() {
         ahrs = new AHRS(I2C.Port.kOnboard);
         System.out.println("Hello World");
         System.out.println(Utility.getFPGATime());
 
-//        talonSrx = new TalonSRX(60);
+        talonSrx = new TalonSRX(60);
     }
 
     @Override
     public void teleopPeriodic() {
         System.out.println(ahrs.getAngle());
-//        talonSrx.set(1.0);
+        talonSrx.set(ControlMode.PercentOutput, 1.0);
     }
 }