+++ src/main/groovy/jaci/openrio/gradle/sim/JavaSimulationTask.groovy
@@ -66,5 +66,6 @@ class JavaSimulationTask extends ExternalLaunchTask {
         persist = true  // So if we crash instantly you can still see the output
         //launch(java, "-Djava.library.path=${ldpath}", "-jar", jar.archivePath.toString())
         // TODO: Add some kind of subsystem here so we can launch externally. It should watch for a stopped build or something
+        println "Simulation is not yet implemented!"
     }
 }